Added assigned incomes deleted categories to other incomes

+++ App/Models/IncomeCategory.php
@@ -60,18 +60,36 @@ class IncomeCategory extends BudgetCategory
 
     public function delete()
     {
-        foreach ($this->categoryToDelete as $category) {
-            $sql = 'DELETE FROM incomes_category_assigned_to_users 
-                    WHERE id = :idCategory';
+        foreach ($this->categoryToDelete as $categoryId) {
+            if (static::assignDeletedCategoriesIncomesToOtherIncomes($categoryId)) {
+                $sql = 'DELETE FROM incomes_category_assigned_to_users 
+                        WHERE id = :idCategory';
 
-            $db = static::getDataBase();
-            $query = $db->prepare($sql);
+                $db = static::getDataBase();
+                $query = $db->prepare($sql);
 
-            $query->bindValue(':idCategory', $category, PDO::PARAM_INT);
-            if (!($query->execute())) {
+                $query->bindValue(':idCategory', $categoryId, PDO::PARAM_INT);
+                if (!($query->execute())) {
+                    return false;
+                }
+            } else {
                 return false;
             }
         }
         return true;
+    }
+
+    protected static function assignDeletedCategoriesIncomesToOtherIncomes($categoryId)
+    {
+        $sql = 'UPDATE incomes
+                SET income_category_assigned_to_user_id = (SELECT id FROM incomes_category_assigned_to_users WHERE name = "Inne przychody" AND user_id = :userId)
+                WHERE income_category_assigned_to_user_id = :idDeletedCategory';
+
+        $db = static::getDataBase();
+        $query = $db->prepare($sql);
+
+        $query->bindValue(':userId', $_SESSION['userId'], PDO::PARAM_INT);
+        $query->bindValue(':idDeletedCategory', $categoryId, PDO::PARAM_INT);
+        return $query->execute();
     }
 }
